% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_area.R
\name{calculate_area}
\alias{calculate_area}
\title{Calculate area}
\usage{
calculate_area(fname)
}
\description{
Function to calculate area from a dataframe that contains 4 columns:
\enumerate{
\item An X column: this should have units metres and be empty apart from the
first entry.
\item A Y column: this should have units metres and be empty apart from the
first entry.
\item A BEARING column, formatted as: HHH°MM'SS"
\item A LENGTH column
}
}
